Add Carte::places and gamedata::_currentPlace. These variables will be used to manage the place where the player is, and will replace SHARED_TRIGGER change_place, Individu::IndiceLieu, Individu::LieuVillage, etc.

+++ src/ElementsCarte/ElementsCarte.h
@@ -111,7 +111,7 @@ class Element_Carte
 		int Id              = -1;
 		string Type         = "DEFAUT";
 		string Liste        = "DEFAUT";
-		short Diplomatie    = 0;
+        int Diplomatie    = 0;
         double lifetime     = -1;
 
 	private:
